refactor: so filter logic from message matcher to filter abstract class
