+++ apps/worker/src/app/workflow/usecases/message-matcher/message-matcher.usecase.spec.ts
@@ -1053,7 +1053,7 @@ describe('Message filter matcher', function () {
     it('should add a passed condition', () => {
       const result = MessageMatcher.sumFilters(
         {
-          stepFilters: [],
+          filters: [],
           failedFilters: [],
           passedFilters: ['payload'],
         },
@@ -1069,14 +1069,14 @@ describe('Message filter matcher', function () {
 
       expect(result.passedFilters).to.contain('payload');
       expect(result.passedFilters.length).to.eq(1);
-      expect(result.stepFilters.length).to.eq(1);
-      expect(result.stepFilters).to.contain('payload');
+      expect(result.filters.length).to.eq(1);
+      expect(result.filters).to.contain('payload');
     });
 
     it('should add a failed condition', () => {
       const result = MessageMatcher.sumFilters(
         {
-          stepFilters: [],
+          filters: [],
           failedFilters: [],
           passedFilters: [],
         },
@@ -1092,14 +1092,14 @@ describe('Message filter matcher', function () {
 
       expect(result.failedFilters).to.contain('payload');
       expect(result.failedFilters.length).to.eq(1);
-      expect(result.stepFilters.length).to.eq(1);
-      expect(result.stepFilters).to.contain('payload');
+      expect(result.filters.length).to.eq(1);
+      expect(result.filters).to.contain('payload');
     });
 
     it('should add online for both cases of online', () => {
       let result = MessageMatcher.sumFilters(
         {
-          stepFilters: [],
+          filters: [],
           failedFilters: [],
           passedFilters: [],
         },
@@ -1115,12 +1115,12 @@ describe('Message filter matcher', function () {
 
       expect(result.passedFilters).to.contain('online');
       expect(result.passedFilters.length).to.eq(1);
-      expect(result.stepFilters.length).to.eq(1);
-      expect(result.stepFilters).to.contain('online');
+      expect(result.filters.length).to.eq(1);
+      expect(result.filters).to.contain('online');
 
       result = MessageMatcher.sumFilters(
         {
-          stepFilters: [],
+          filters: [],
           failedFilters: [],
           passedFilters: [],
         },
@@ -1136,8 +1136,8 @@ describe('Message filter matcher', function () {
 
       expect(result.passedFilters).to.contain('online');
       expect(result.passedFilters.length).to.eq(1);
-      expect(result.stepFilters.length).to.eq(1);
-      expect(result.stepFilters).to.contain('online');
+      expect(result.filters.length).to.eq(1);
+      expect(result.filters).to.contain('online');
     });
   });
 });
