--- apps/worker/src/app/workflow/usecases/message-matcher/types.ts
@@ -1,8 +1,0 @@
-import type { ITriggerPayload } from '@novu/shared';
-import type { SubscriberEntity } from '@novu/dal';
-
-export interface IFilterVariables {
-  payload?: ITriggerPayload;
-  subscriber?: SubscriberEntity;
-  webhook?: Record<string, unknown>;
-}
