+++ packages/application-generic/src/index.ts
@@ -7,4 +7,6 @@ export * from './logging/index';
 export * from './usecases';
 export * from './instrumentation/index';
 export * from './utils/subscriber';
+export * from './utils/filter';
+export * from './utils/filter-processing-details';
 export * from './resilience';
